- some fixes for Visual Studio
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@8999 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/src/ParametricCurves.cpp
@@ -218,7 +218,7 @@ void computeBezierPoints(const vector<Coord> &controlPoints, vector<Coord> &curv
     #pragma omp parallel for
 #endif
 
-    for (unsigned int i = 0 ; i < nbCurvePoints ; ++i) {
+    for (int i = 0 ; i < static_cast<int>(nbCurvePoints) ; ++i) {
       float curStep = i * h;
       curvePoints[i] = computeBezierPoint(controlPoints, curStep);
     }
